+++ packages/main/package.json
@@ -18,9 +18,10 @@
     "serve": "wc-dev serve",
     "build": "wc-dev build",
     "test": "wc-dev test",
-    "scope-test-pages": "wc-dev scopeTestPages",
     "create-ui5-element": "wc-create-ui5-element",
-    "prepublishOnly": "npm run build"
+    "prepublishOnly": "npm run build",
+    "scope-test-pages": "wc-dev scopeTestPages",
+    "scope-lint": "wc-dev scopeLint"
   },
   "repository": {
     "type": "git",
